+++ modules/main/src/main/java/com/nisovin/shopkeepers/text/TextBuilder.java
@@ -261,21 +261,19 @@ public abstract class TextBuilder extends AbstractText {
 	 * Creates a new {@link TextBuilder} with the specified hover event and sets it as
 	 * {@link #child(Text) child} Text.
 	 * 
-	 * @param action
-	 *            the hover event action, not <code>null</code>
-	 * @param value
-	 *            the hover event value, not <code>null</code>
+	 * @param content
+	 *            the hover event content, not <code>null</code>
 	 * @return the new {@link TextBuilder}
 	 */
-	public TextBuilder childHoverEvent(HoverEventText.Action action, Text value) {
-		return this.child(Text.hoverEvent(action, value));
+	public TextBuilder childHoverEvent(HoverEventText.Content content) {
+		return this.child(Text.hoverEvent(content));
 	}
 
 	/**
 	 * Creates a new {@link TextBuilder} with the specified hover text and sets it as
 	 * {@link #child(Text) child} Text.
 	 * <p>
-	 * This is a shortcut for the corresponding {@link #hoverEvent(HoverEventText.Action, Text)}.
+	 * This is a shortcut for the corresponding {@link #childHoverEvent(HoverEventText.Content)}.
 	 * 
 	 * @param hoverText
 	 *            the hover text, not <code>null</code>
@@ -423,21 +421,19 @@ public abstract class TextBuilder extends AbstractText {
 	 * Creates a new {@link TextBuilder} with the specified hover event and sets it as
 	 * {@link #next(Text) next} Text.
 	 * 
-	 * @param action
-	 *            the hover event action, not <code>null</code>
-	 * @param value
-	 *            the hover event value, not <code>null</code>
+	 * @param content
+	 *            the hover event content, not <code>null</code>
 	 * @return the new {@link TextBuilder}
 	 */
-	public TextBuilder hoverEvent(HoverEventText.Action action, Text value) {
-		return this.next(Text.hoverEvent(action, value));
+	public TextBuilder hoverEvent(HoverEventText.Content content) {
+		return this.next(Text.hoverEvent(content));
 	}
 
 	/**
 	 * Creates a new {@link TextBuilder} with the specified hover text and sets it as
 	 * {@link #next(Text) next} Text.
 	 * <p>
-	 * This is a shortcut for the corresponding {@link #hoverEvent(HoverEventText.Action, Text)}.
+	 * This is a shortcut for the corresponding {@link #hoverEvent(HoverEventText.Content)}.
 	 * 
 	 * @param hoverText
 	 *            the hover text, not <code>null</code>